pyproject.toml
[tool.poetry]
name = "psy-diary"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^23.12.1"
python-telegram-bot = "^20.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
